chat done + search halfway

+++ app/src/main/java/ge/nnasaridze/messengerapp/scenes/chat/presentation/recycler/RecyclerMessageEntity.kt
@@ -2,6 +2,6 @@ package ge.nnasaridze.messengerapp.scenes.chat.presentation.recycler
 
 data class RecyclerMessageEntity(
     var text: String,
-    var timestamp: Int,
+    var timestamp: Long,
     var sent: Boolean,
 )
